services/guardian/pyproject.toml
[tool.poetry]
name = "clestiq-shield-guardian"
version = "1.0.0"
description = "Clestiq Shield Guardian - Output Validation Agent"
authors = ["Admin <vasubhut157@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic-settings = "^2.1.0"
pydantic = "^2.0.0"
structlog = "^24.1.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-instrumentation-logging = "^0.43b0"
python-json-logger = "^2.0.7"
langchain = "1.0.0"
langgraph = "1.0.0"
langchain-google-vertexai = "3.1.0"
bleach = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
httpx = ">=0.28.0,<1.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
